added tests, added README
