+++ README.txt
@@ -1,0 +1,17 @@
+Introduction
+
+This is a simple Rock-Paper-Scissors application. You play against Curb (Server). If it doesn't work, the system makes a random move instead of it.
+The result is displayed on the screen.
+
+Technical stack
+
+-Java 11
+-SPRING BOOT
+-MAVEN
+-Lombok
+-JUnit4
+-Thymeleaf
+
+Technical description
+
+The application comprises several packages, split by SOLID principles. View is written using Thymeleaf. Main class - RockPaperScissorsApplication.
