+++ src/main/java/com/ua/game/demo/service/GameEngineService.java
@@ -19,6 +19,8 @@ public class GameEngineService {
     private GameRulesService service;
     @Autowired
     private CurbController controller;
+    @Autowired
+    private RandomThrowService randomThrowService;
 
     public String findWinner(String userChoice) {
         log.info("User's choice: {}", userChoice);
@@ -26,26 +28,12 @@ public class GameEngineService {
         try {
             systemMove = controller.getResponse();
         } catch (IOException e) {
-            systemMove = calculateRandomThrow();
+            systemMove = randomThrowService.calculateRandomThrow();
         }
         if (userChoice.equalsIgnoreCase(systemMove)) {
             return IT_S_A_DRAW;
         } else {
             return service.isPlayerWin(userChoice, systemMove) ? YOU_WIN : YOU_LOSE;
         }
-    }
-
-    private String calculateRandomThrow() {
-        int rand = (int) (Math.random() * 3);
-
-        String move = "";
-        if (rand == 0) {
-            move = "rock";
-        } else if (rand == 1) {
-            move = "paper";
-        } else {
-            move = "scissors";
-        }
-        return move;
     }
 }
